Code-gen native lexical/attr access as ref-taking.
This makes a number of things work that never have before (of note, ++
on a native variable). However, various things need updating to know
about native refs (so there are quite some spectest regressions), and
the many, many possible code-gen optimizations are not yet in place,
let alone any of the cleverer VM-level ones.